Move Sport to :common
