+++ app/src/main/java/com/ricknout/rugbyranker/db/WorldRugbyRankingDao.kt
@@ -5,7 +5,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.ricknout.rugbyranker.vo.Sport
+import com.ricknout.rugbyranker.common.vo.Sport
 import com.ricknout.rugbyranker.vo.WorldRugbyRanking
 
 @Dao
