+++ app/src/main/java/com/ricknout/rugbyranker/ui/matches/MatchesViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.ViewModel
 import com.ricknout.rugbyranker.common.livedata.Event
 import com.ricknout.rugbyranker.repository.RugbyRankerRepository
 import com.ricknout.rugbyranker.vo.MatchStatus
-import com.ricknout.rugbyranker.vo.Sport
+import com.ricknout.rugbyranker.common.vo.Sport
 import com.ricknout.rugbyranker.vo.WorldRugbyMatch
 import com.ricknout.rugbyranker.work.RugbyRankerWorkManager
