+++ app/src/main/java/com/ricknout/rugbyranker/ui/matches/WomensCompleteMatchesViewModel.kt
@@ -2,7 +2,7 @@ package com.ricknout.rugbyranker.ui.matches
 
 import com.ricknout.rugbyranker.repository.RugbyRankerRepository
 import com.ricknout.rugbyranker.vo.MatchStatus
-import com.ricknout.rugbyranker.vo.Sport
+import com.ricknout.rugbyranker.common.vo.Sport
 import com.ricknout.rugbyranker.work.RugbyRankerWorkManager
 import javax.inject.Inject
