+++ app/src/main/java/com/ricknout/rugbyranker/ui/rankings/MensRankingsViewModel.kt
@@ -1,7 +1,7 @@
 package com.ricknout.rugbyranker.ui.rankings
 
 import com.ricknout.rugbyranker.repository.RugbyRankerRepository
-import com.ricknout.rugbyranker.vo.Sport
+import com.ricknout.rugbyranker.common.vo.Sport
 import com.ricknout.rugbyranker.work.RugbyRankerWorkManager
 import javax.inject.Inject
