+++ app/src/main/java/com/ricknout/rugbyranker/ui/rankings/RankingsFragment.kt
@@ -19,7 +19,7 @@ import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.work.WorkInfo.State
 import com.google.android.material.snackbar.Snackbar
 import com.ricknout.rugbyranker.common.util.doIfVisibleToUser
-import com.ricknout.rugbyranker.vo.Sport
+import com.ricknout.rugbyranker.common.vo.Sport
 import kotlinx.android.synthetic.main.fragment_rankings.*
 
 class RankingsFragment : DaggerFragment() {
