+++ app/src/main/java/com/ricknout/rugbyranker/work/MensWorldRugbyRankingsWorker.kt
@@ -3,7 +3,7 @@ package com.ricknout.rugbyranker.work
 import android.content.Context
 import androidx.work.WorkerParameters
 import com.ricknout.rugbyranker.repository.RugbyRankerRepository
-import com.ricknout.rugbyranker.vo.Sport
+import com.ricknout.rugbyranker.common.vo.Sport
 
 class MensWorldRugbyRankingsWorker(
         context: Context,
