+++ app/src/main/java/com/ricknout/rugbyranker/work/WorldRugbyMatchesWorker.kt
@@ -5,7 +5,7 @@ import androidx.work.Worker
 import androidx.work.WorkerParameters
 import com.ricknout.rugbyranker.repository.RugbyRankerRepository
 import com.ricknout.rugbyranker.vo.MatchStatus
-import com.ricknout.rugbyranker.vo.Sport
+import com.ricknout.rugbyranker.common.vo.Sport
 
 open class WorldRugbyMatchesWorker(
         context: Context,
