+++ app/src/main/java/com/ricknout/rugbyranker/work/WorldRugbyRankingsWorker.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import androidx.work.Worker
 import androidx.work.WorkerParameters
 import com.ricknout.rugbyranker.repository.RugbyRankerRepository
-import com.ricknout.rugbyranker.vo.Sport
+import com.ricknout.rugbyranker.common.vo.Sport
 
 open class WorldRugbyRankingsWorker(
         context: Context,
