+++ app/src/test/java/com/ricknout/rugbyranker/vo/WorldRugbyRankingTest.kt
@@ -1,5 +1,6 @@
 package com.ricknout.rugbyranker.vo
 
+import com.ricknout.rugbyranker.common.vo.Sport
 import org.junit.Test
 import org.junit.Assert.assertEquals
