+++ common/src/main/java/com/ricknout/rugbyranker/common/vo/Sport.kt
@@ -1,4 +1,4 @@
-package com.ricknout.rugbyranker.vo
+package com.ricknout.rugbyranker.common.vo
 
 enum class Sport {
     MENS,
